cigar_filter folder added
cigar filter contains cigar_filter.py that filters out soft clipped
reads. It allows reads that overlap the edges of the contig ends to have
soft clipping. It also permits reads with soft clipping of 4 or fewer bases of
or end within the body of the alignment that is mapped to the contig.