consensus_caller added
Wrote pysam consensus caller pysam_consensus.py which uses functions from get_insert.py
get_insert.py is built around parsing out the insert cigar strings while processing the
reads in pileupcolumns. by using pileupread.indel  and pileupcolumn.pos it is possible to determine
the position of the insertion in the raw read and return it. Once returned insertion events can be
tabulated and compared to other possible bases for that position in the contig, if it is the most
abundant character and above a pre-determined cutoff, then it added to the consensus sequence. Deletion
events are tabulated on the fly the same as a single bp. If the map is well aligned, processing deletions
in this way should not be a problem. pysam_consensus.py pads uncovered characters between start of contig and
first mapped position with character P. It reports positions for which there is no single base above the cutoff
as N, and it provides no internal padding for unmapped regions.